Fix?(__init__.py, hogwarts.py)
Put error handling back in __init__.py, as it is global to the app and affects
all blueprints. Put static file serving in __init__.py as well, in order to
avoid the use of redirects which 1) are inefficient and 2) expose our (silly)
internal filenames to the outside world.
