+++ getpost/__init__.py
@@ -1,4 +1,5 @@
-from flask import Flask
+from flask import Flask, render_template, send_from_directory
+from os.path import join, dirname, abspath
 
 from .desk.hogwarts import hogwarts_blueprint
 from .desk.boats import boats_blueprint
@@ -18,5 +19,38 @@ def create_app(config_obj):
     app.register_blueprint(owls_blueprint)
     app.register_blueprint(parcels_blueprint)
     app.register_blueprint(wizards_blueprint)
+
+    @app.errorhandler(500)
+    def internal(error):
+        desc = 'Uh oh! Something went wrong.'
+        return render_template(
+            'voldemort.html', status=500, description=desc
+        ), 500
+
+
+    @app.errorhandler(404)
+    def not_found(error):
+        desc = 'This page does not exist.'
+        return render_template(
+            'voldemort.html', status=404, description=desc
+        ), 404
+
+    static_directory = join(abspath(dirname(__file__)), 'static/')
+
+    css_codenames = {'template.css': 'maraudersmap.css',
+                     'home.css': 'hogwarts.css',
+                     'signup.css': 'boats.css'}
+    @app.route('/css/<path:path>')
+    def send_css(path):
+        if path in css_codenames:
+            path = css_codenames[path]
+        return send_from_directory(join(static_directory, 'css/'), path)
+
+    js_codenames = {'signup.js': 'boats.js'}
+    @app.route('/js/<path:path>')
+    def send_js(path):
+        if path in js_codenames:
+            path = js_codenames[path]
+        return send_from_directory(join(static_directory, 'js/'), path)
 
     return app
